requirements.txt
# Core dependencies
pyyaml>=6.0
requests>=2.31.0
numpy>=1.24.0

# Vector database
chromadb>=0.4.0

# Text processing
langchain>=0.0.350
langchain-community>=0.0.10

# Alternative vector stores (optional)
faiss-cpu>=1.7.4

# CLI and utilities
click>=8.1.0
rich>=13.0.0
tqdm>=4.65.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0